[Prover] Extract on-chain keyless config logic.
