+++ prover-service/src/lib.rs
@@ -9,10 +9,10 @@ pub mod groth16_vk;
 pub mod handlers;
 pub mod input_processing;
 pub mod jwk_fetching;
+pub mod keyless_config;
 pub mod load_vk;
 pub mod metrics;
 pub mod prover_config;
-pub mod prover_key;
 pub mod prover_state;
 pub mod proving;
 pub mod training_wheels;
