+++ prover-service/src/main.rs
@@ -10,7 +10,6 @@ use axum::{
 use clap::Parser;
 use http::Method;
 use prover_service::prover_config::ProverServiceConfig;
-use prover_service::prover_key::TrainingWheelsKeyPair;
 use prover_service::{prover_state::*, *};
 use std::{net::SocketAddr, sync::Arc, time::Duration};
 use tower::ServiceBuilder;
